V3.2.0 (#184)
* Fix bug when journal entry amount is 0

A journal entry amount of "0" would always results in a transaction type of "Debit". Add the ability to indicate which journal entry is Debit and which is Credit when journaling an amount of "0".

* Allow unsetting email address in vendor

Allow clearing a previously-set email address in the Vendor.  Previously, once you set an email address, you could never remove an email address.

* Write the "start" XML attribute

Write the "start" XML attribute when paging through a list.

* Allow using record number when updating customer or vendor

Previously, the library only supported using the vendor ID or customer ID when updating a vendor or customer.  Allow additionally using the record number to update.

* Allow using ReadMore for reports

Listing stored report data requires using the ReadMore function. Add a report ID to the ReadMore function.

* Fix the unit tests

After augmenting VendorUpdate, CustomerUpdate, and ReadMore to allow record ID (for the former two) and report ID (for the latter), fix the corresponding unit tests.

* Update JournalEntryLineCreate.php

Fix bug in setting transaction amount

* Fix for issue 168

* added tax entries for bill

* updated referenced versions for travis and composer to 8.1

* added description to new methods

* updated set description for primaryEmailAddress

* bump version 3.1.0 to 4.0.0

* corrected version bump from 4.0.0 to 3.2.0

* updated supported versions of php

* remove .travis.yml

* updated doxygen related files

Co-authored-by: Ben Hardin <31776675+bch36@users.noreply.github.com>